+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 2.5.6 (Unreleased)
+## 2.5.6-preview.1 (2026-09-25)
 **Features**
 
 **Bug Fixes**